Control.CollectErrors

Monad for collecting errors in expressions

data CollectErrors es v

A wrapper around values which can accommodate a list of (potential) errors that have (maybe) occurred during the computation of a value. A value may be missing, leaving only the error(s).

Such error collection allows one to write expressions with partial functions (ie functions that fail for some inputs) instead of branching after each application of such function. Dealing with the errors can be moved outside the expression. If the error data contain enough information, their list can be used to trace the source of the errors.

Constructors

CollectErrors

Fields

getMaybeValue :: Maybe v

getErrors :: es

noErrors :: Monoid es => v -> CollectErrors es v

Wrap a pure value in a CollectErrors record

noValue :: es -> CollectErrors es v

Make CollectErrors record with no value, only errors.

prependErrors :: Monoid es => es -> CollectErrors es v -> CollectErrors es v

Add further errors into a CollectErrors record.

getValueIfNoError :: (Monoid es, Eq es) => CollectErrors es v -> (v -> t) -> (es -> t) -> t

A safe way to get a value out of the CollectErrors wrapper.

getValueOrThrowErrors :: (Show es, Monoid es, Eq es) => CollectErrors es v -> v

An unsafe way to get a value out of the CollectErrors wrapper.

lift1 :: (a -> b) -> CollectErrors es a -> CollectErrors es b

Add error collection support to an unary operation.

lift2 :: Monoid es => (a -> b -> c) -> CollectErrors es a -> CollectErrors es b -> CollectErrors es c

Add error collection support to a binary operation.

unlift2first :: Monoid es => (CollectErrors es a -> CollectErrors es b -> t) -> a -> CollectErrors es b -> t

An utility function for easily defining a binary function with only the second parameter collecting errors, if we have an analogous function that requires both parameters to collect errors.

unlift2second :: Monoid es => (CollectErrors es a -> CollectErrors es b -> t) -> CollectErrors es a -> b -> t

An utility function for easily defining a binary function with only the first parameter collecting errors, if we have an analogous function that requires both parameters to collect errors.

Tools for avoiding CollectErrors(CollectErrors t)

ensureCE :: CanEnsureCE es v => v -> EnsureCE es v

Translate a value of a type a to a value of a type CollectErrors es a except when a already is a CollectErrors type, in which case the value is left as is.

type CanEnsureCE es v = CanEnsureTypeOp (CollectErrors es) v

type EnsureCE es v = EnsureTypeOp (CollectErrors es) v

type WithoutCE es v = RemoveTypeOp (CollectErrors es) v

Apply CollectErrors to a type except when the type already is a CollectErrors type.

Remove CollectErrors wrapper from a type if it is there.

lift1ensureCE :: (Monoid es, CanEnsureCE es b) => (a -> b) -> CollectErrors es a -> EnsureCE es b

Add error collection support to a binary function whose result may already have collected errors.

lift2ensureCE :: (Monoid es, CanEnsureCE es c) => (a -> b -> c) -> CollectErrors es a -> CollectErrors es b -> EnsureCE es c

Add error collection support to a binary function whose result may already have collected errors.