Control.EnsureTypeOp

Documentation

type EnsureTypeOp f a = f (RemoveTypeOp f a)

A mechanism for applying a type operator to a type except when the type already has that operator.

type family RemoveTypeOp f a where ...

Equations

RemoveTypeOp f (f a) = a

RemoveTypeOp f a = a

class CanEnsureTypeOp f a where

A mechanism for translating a value of a type a to a value of a type f a except when a = f b for some b, in which case the value is left as is.

Minimal complete definition

ensureTypeOp

Methods

ensureTypeOp :: a -> EnsureTypeOp f a