Numeric.CollectErrors

Description

A type of numeric errors to be collected.

Type of numeric errors

data ErrorCertaintyLevel

Constructors

ErrorCertain

ErrorPotential

data NumError

Constructors

DivByZero

OutOfRange String

NumError String

type NumErrors = [(ErrorCertaintyLevel, NumError)]

sample_NumErrors :: Maybe [(ErrorCertaintyLevel, NumError)]

Specialisation to numeric errors

type CN = CollectErrors NumErrors

hasCertainError :: CanTestErrorsCertain es => es -> Bool

hasCertainErrorCN :: CN v -> Bool

noValueCN :: NumErrors -> CN v

noValueNumErrorCertainCN :: NumError -> CN v

Construct an empty wrapper indicating that given error has certainly occurred.

noValueNumErrorPotentialCN :: NumError -> CN v

Construct an empty wrapper indicating that given error may have occurred.

getMaybeValueCN :: CN v -> Maybe v

getErrorsCN :: CN v -> NumErrors

prependErrorsCN :: NumErrors -> CN v -> CN v

type CanEnsureCN = CanEnsureCE NumErrors

type EnsureCN a = EnsureCE NumErrors a

type EnsureNoCN a = EnsureNoCE NumErrors a

ensureCN :: CanEnsureCN v => v -> EnsureCN v

Translate a value of a type a to a value of a type CollectNumErrors a except when a already is a CollectNumErrors type, in which case the value is left as is.

deEnsureCN :: CanEnsureCN v => EnsureCN v -> Either NumErrors v

Translate a value of a type EnsureCN es a to a, throwing an exception if there was an error. If a is a CollectNumErrors type, then this is just an identity.

ensureNoCN :: CanEnsureCN v => v -> (Maybe (EnsureNoCN v), NumErrors)

Translate a value of a type a to a value of a type CollectNumErrors a except when a already is a CollectNumErrors type, in which case the value is left as is.

noValueECN :: CanEnsureCN v => Maybe v -> NumErrors -> EnsureCN v

prependErrorsECN :: CanEnsureCN v => Maybe v -> NumErrors -> EnsureCN v -> EnsureCN v

noValueNumErrorCertainECN :: CanEnsureCN v => Maybe v -> NumError -> EnsureCN v

Construct an empty wrapper indicating that given error has certainly occurred.

noValueNumErrorPotentialECN :: CanEnsureCN v => Maybe v -> NumError -> EnsureCN v

Construct an empty wrapper indicating that given error may have occurred.

type CanExtractCN f = CanExtractCE NumErrors f

extractCN :: (CanEnsureCN c, CanExtractCN f) => f c -> CN (f (EnsureNoCN c))

More compact synonyms

cn :: CanEnsureCN v => v -> EnsureCN v

Wrap a value in the CollectNumErrors wrapper.

deCN :: CanEnsureCN v => EnsureCN v -> Either NumErrors v

An unsafe way to get a value out of the CollectNumErrors wrapper.

(~!) :: (CanEnsureCN v, Show v) => v -> EnsureNoCN v

An unsafe way to get a value out of the CollectNumErrors wrapper.