Numeric.MixedTypes.AddSub

Description

Addition

type CanAdd t1 t2 = (CanAddAsymmetric t1 t2, CanAddAsymmetric t2 t1, AddType t1 t2 ~ AddType t2 t1)

class CanAddAsymmetric t1 t2 where

A replacement for Prelude's +. If t1 = t2 and Num t1, then one can use the default implementation to mirror Prelude's +.

Associated Types

type AddType t1 t2

Methods

add :: t1 -> t2 -> AddType t1 t2

add :: (AddType t1 t2 ~ t1, t1 ~ t2, Num t1) => t1 -> t1 -> t1

type CanAddThis t1 t2 = (CanAdd t1 t2, AddType t1 t2 ~ t1)

type CanAddSameType t = CanAddThis t t

(+) :: CanAddAsymmetric t1 t2 => t1 -> t2 -> AddType t1 t2 infixl 6

sum :: (CanAddSameType t, ConvertibleExactly Integer t) => [t] -> t

Tests

specCanAdd :: (CanAddXX t1 t1, CanAddXX t1 t2, CanAddXX t1 t3, CanAddXX t2 t3, CanAddXX t1 (AddType t2 t3), CanAddXX (AddType t1 t2) t3, ConvertibleExactly Integer t1, CanTestPosNeg t1, HasEqCertainly (AddType t1 (AddType t2 t3)) (AddType (AddType t1 t2) t3)) => T t1 -> T t2 -> T t3 -> Spec

HSpec properties that each implementation of CanAdd should satisfy.

specCanAddNotMixed :: (CanAddXX t t, CanAddXX t (AddType t t), ConvertibleExactly Integer t, CanTestPosNeg t) => T t -> Spec

HSpec properties that each implementation of CanAdd should satisfy.

specCanAddSameType :: (ConvertibleExactly Integer t, Show t, HasEqCertainly t t, CanAddSameType t) => T t -> Spec

HSpec properties that each implementation of CanAddSameType should satisfy.

type CanAddX t1 t2 = (CanAdd t1 t2, Show t1, Arbitrary t1, Show t2, Arbitrary t2, Show (AddType t1 t2), HasEqCertainly t1 (AddType t1 t2), HasEqCertainly t2 (AddType t1 t2), HasEqCertainly (AddType t1 t2) (AddType t1 t2), HasOrderCertainly t1 (AddType t1 t2), HasOrderCertainly t2 (AddType t1 t2), HasOrderCertainly (AddType t1 t2) (AddType t1 t2))

Compound type constraint useful for test definition.

type CanAddXX t1 t2 = (CanAddX t1 t2, HasEqCertainly (AddType t1 t2) (AddType t2 t1))

Compound type constraint useful for test definition.

Subtraction

class CanSub t1 t2 where

A replacement for Prelude's binary -.

If CanNeg t2 and CanAdd t1 (NegType t2), then one can use the default implementation via a-b = a + (-b).

Associated Types

type SubType t1 t2

Methods

sub :: t1 -> t2 -> SubType t1 t2

sub :: (SubType t1 t2 ~ AddType t1 (NegType t2), CanNeg t2, CanAdd t1 (NegType t2)) => t1 -> t2 -> SubType t1 t2

type CanSubThis t1 t2 = (CanSub t1 t2, SubType t1 t2 ~ t1)

type CanSubSameType t = CanSubThis t t

(-) :: CanSub t1 t2 => t1 -> t2 -> SubType t1 t2 infixl 6

Tests

specCanSub :: (CanSubX t1 t1, CanSubX t1 t2, CanNeg t2, CanAdd t1 (NegType t2), HasEqCertainly (SubType t1 t2) (AddType t1 (NegType t2)), Show (AddType t1 (NegType t2)), ConvertibleExactly Integer t1) => T t1 -> T t2 -> Spec

HSpec properties that each implementation of CanSub should satisfy.

specCanSubNotMixed :: (CanSubX t t, CanSubX t (SubType t t), CanNeg t, CanAdd t (NegType t), Show (AddType t (NegType t)), HasEqCertainly (SubType t t) (AddType t (NegType t)), ConvertibleExactly Integer t) => T t -> Spec

HSpec properties that each implementation of CanSub should satisfy.

type CanSubX t1 t2 = (CanSub t1 t2, HasEqCertainly t1 (SubType t1 t2), CanAddXX t1 t2, Show (SubType t1 t2))

Compound type constraint useful for test definition.