Numeric.MixedTypes.Bool

Description

Documentation

type IsBool t = (HasBools t, CanNegSameType t, CanAndOrSameType t)

A type constraint synonym that stipulates that the type behaves very much like Bool, except it does not necessarily satisfy the law of excluded middle, which means that the type can contain a "do-not-know" value.

Examples: Bool, Maybe Bool, Maybe (Maybe Bool), CollectErrors Bool

specIsBool :: (IsBool t, CanTestCertainly t, Show t, Serial IO t) => T t -> Spec

HSpec properties that each implementation of IsBool should satisfy.

Conversion to/from Bool

type HasBools t = ConvertibleExactly Bool t

class HasBools t => CanTestCertainly t where

Tests for truth or falsity. Beware, when isCertainlyTrue returns False, it does not mean that the proposition is false. It usually means that we failed to prove the proposition.

Minimal complete definition

isCertainlyTrue, isCertainlyFalse

Methods

isCertainlyTrue :: t -> Bool

isCertainlyFalse :: t -> Bool

specCanTestCertainly :: CanTestCertainly t => T t -> Spec

HSpec properties that each implementation of CanTestCertainly should satisfy.

type CanTestCertainlyX t = (CanTestCertainly t, Show t, Serial IO t)

Compound type constraint useful for test definition.

isNotTrue :: CanTestCertainly t => t -> Bool

isNotFalse :: CanTestCertainly t => t -> Bool

stronglyImplies :: (CanTestCertainly t1, CanTestCertainly t2) => t1 -> t2 -> Bool

If l is certainly True, then r is also certainly True.

stronglyEquivalentTo :: (CanTestCertainly t1, CanTestCertainly t2) => t1 -> t2 -> Bool

weaklyImplies :: (CanTestCertainly t1, CanTestCertainly t2) => t1 -> t2 -> Bool

If l is certainly True, then r is not certainly False.

weaklyEquivalentTo :: (CanTestCertainly t1, CanTestCertainly t2) => t1 -> t2 -> Bool

Negation

class CanNeg t where

This is negation is both the numeric negation as well as the Boolean negation. Example of non-standard Boolean negation:

 negate (Just True) = Just False
 

Minimal complete definition

negate

Associated Types

type NegType t

Methods

negate :: t -> NegType t

not :: CanNeg t => t -> NegType t

A synonym of negate.

type CanNegSameType t = (CanNeg t, NegType t ~ t)

Tests

specCanNegBool :: (Show t, Show (NegType (NegType t)), Serial IO t, CanTestCertainly t, CanTestCertainly (NegType t), CanTestCertainly (NegType (NegType t)), CanNeg t, CanNeg (NegType t)) => T t -> Spec

HSpec properties that each Boolean implementation of CanNeg should satisfy.

And and or

type CanAndOr t1 t2 = (CanAndOrAsymmetric t1 t2, CanAndOrAsymmetric t2 t1, AndOrType t1 t2 ~ AndOrType t2 t1)

class CanAndOrAsymmetric t1 t2 where

Binary logical and and or for generalised Booleans. For example:

 (Just True) && False = Just False
 (Just (Just True)) || False = (Just (Just True))
 

Minimal complete definition

and2, or2

Associated Types

type AndOrType t1 t2

Methods

and2 :: t1 -> t2 -> AndOrType t1 t2

or2 :: t1 -> t2 -> AndOrType t1 t2

(&&) :: CanAndOrAsymmetric a b => a -> b -> AndOrType a b infixr 3

A synonym of and2.

(||) :: CanAndOrAsymmetric a b => a -> b -> AndOrType a b infixr 2

A synonym of or2.

type CanAndOrWith t1 t2 = (CanAndOr t1 t2, AndOrType t1 t2 ~ t1)

type CanAndOrSameType t = CanAndOrWith t t

and :: (CanAndOrSameType t, CanTestCertainly t) => [t] -> t

or :: (CanAndOrSameType t, CanTestCertainly t) => [t] -> t

Tests

specCanAndOr :: (Show t1, Show t2, Show t3, Show (AndOrType t1 t1), Show (AndOrType t1 t2), Show (AndOrType t2 t1), Show (AndOrType t1 (AndOrType t2 t3)), Show (AndOrType (AndOrType t1 t2) t3), Show (AndOrType (AndOrType t1 t2) (AndOrType t1 t3)), Show (NegType (AndOrType t1 t2)), Show (AndOrType (NegType t1) (NegType t2)), Serial IO t1, Serial IO t2, Serial IO t3, CanTestCertainly t1, CanTestCertainly (AndOrType t1 t1), CanTestCertainly (AndOrType t1 t2), CanTestCertainly (AndOrType t2 t1), CanTestCertainly (AndOrType t1 (AndOrType t2 t3)), CanTestCertainly (AndOrType (AndOrType t1 t2) t3), CanTestCertainly (AndOrType (AndOrType t1 t2) (AndOrType t1 t3)), CanTestCertainly (NegType (AndOrType t1 t2)), CanTestCertainly (AndOrType (NegType t1) (NegType t2)), CanNeg t1, CanNeg t2, CanNeg (AndOrType t1 t2), CanAndOrAsymmetric t1 t1, CanAndOrAsymmetric t1 t2, CanAndOrAsymmetric t1 t3, CanAndOrAsymmetric t1 (AndOrType t2 t3), CanAndOrAsymmetric t2 t1, CanAndOrAsymmetric t2 t3, CanAndOrAsymmetric (AndOrType t1 t2) t3, CanAndOrAsymmetric (AndOrType t1 t2) (AndOrType t1 t3), CanAndOrAsymmetric (NegType t1) (NegType t2)) => T t1 -> T t2 -> T t3 -> Spec

HSpec properties that each implementation of CanAndOr should satisfy.

specCanAndOrNotMixed :: (Show t, Show (AndOrType t t), Show (AndOrType t (AndOrType t t)), Show (AndOrType (AndOrType t t) t), Show (AndOrType (AndOrType t t) (AndOrType t t)), Show (NegType (AndOrType t t)), Show (AndOrType (NegType t) (NegType t)), Serial IO t, CanTestCertainly t, CanTestCertainly (AndOrType t t), CanTestCertainly (AndOrType t (AndOrType t t)), CanTestCertainly (AndOrType (AndOrType t t) t), CanTestCertainly (AndOrType (AndOrType t t) (AndOrType t t)), CanTestCertainly (NegType (AndOrType t t)), CanTestCertainly (AndOrType (NegType t) (NegType t)), CanNeg t, CanNeg (AndOrType t t), CanAndOrAsymmetric t t, CanAndOrAsymmetric t (AndOrType t t), CanAndOrAsymmetric (AndOrType t t) t, CanAndOrAsymmetric (AndOrType t t) (AndOrType t t), CanAndOrAsymmetric (NegType t) (NegType t)) => T t -> Spec

HSpec properties that each implementation of CanAndOr should satisfy.

Orphan instances

ConvertibleExactly Bool Bool

Methods

safeConvertExactly :: Bool -> ConvertResult Bool

ConvertibleExactly Bool t => ConvertibleExactly Bool (Maybe t)

Methods

safeConvertExactly :: Bool -> ConvertResult (Maybe t)