Numeric.MixedTypes.Elementary

Description

Square root

class CanSqrt t where

A replacement for Prelude's sqrt. If Floating t, then one can use the default implementation to mirror Prelude's sqrt.

Associated Types

type SqrtType t

Methods

sqrt :: t -> SqrtType t

sqrt :: (SqrtType t ~ t, Floating t) => t -> t

type CanSqrtSameType t = (CanSqrt t, SqrtType t ~ t)

specCanSqrtReal :: (CanSqrtX t, CanPowX (SqrtType t) Integer, HasEqCertainly t (PowType (SqrtType t) Integer)) => T t -> Spec

HSpec properties that each implementation of CanSqrt should satisfy.

Exp

class CanExp t where

A replacement for Prelude's exp. If Floating t, then one can use the default implementation to mirror Prelude's exp.

Associated Types

type ExpType t

Methods

exp :: t -> ExpType t

exp :: (ExpType t ~ t, Floating t) => t -> t

type CanExpSameType t = (CanExp t, ExpType t ~ t)

specCanExpReal :: CanExpX t => T t -> Spec

HSpec properties that each implementation of CanExp should satisfy.

Log

class CanLog t where

A replacement for Prelude's log. If Floating t, then one can use the default implementation to mirror Prelude's log.

Associated Types

type LogType t

Methods

log :: t -> LogType t

log :: (LogType t ~ t, Floating t) => t -> t

type CanLogSameType t = (CanLog t, LogType t ~ t)

specCanLogReal :: (CanLogX t, CanLogX (DivType Integer t), CanExp t, CanLogX (ExpType t), HasEqCertainly t (LogType (ExpType t))) => T t -> Spec

HSpec properties that each implementation of CanLog should satisfy.

powUsingExpLog :: (CanTestPosNeg t1, CanMulSameType t1, HasIntegers t1, CanTestZero t1, CanRecipSameType t1, CanTestInteger t2, CanTestPosNeg t2, CanMulAsymmetric (LogType t1) t2, CanLog t1, CanExp (MulType (LogType t1) t2), ExpType (MulType (LogType t1) t2) ~ t1) => t1 -> t2 -> ExpType (MulType (LogType t1) t2)

Sine and cosine

class CanSinCos t where

A replacement for Prelude's cos and sin. If Floating t, then one can use the default implementation to mirror Prelude's sin, cos.

Associated Types

type SinCosType t

Methods

cos :: t -> SinCosType t

cos :: (SinCosType t ~ t, Floating t) => t -> t

sin :: t -> SinCosType t

sin :: (SinCosType t ~ t, Floating t) => t -> t

type CanSinCosSameType t = (CanSinCos t, SinCosType t ~ t)

specCanSinCosReal :: CanSinCosX t => T t -> Spec

HSpec properties that each implementation of CanSinCos should satisfy.

Derived partially from http://math.stackexchange.com/questions/1303044/axiomatic-definition-of-sin-and-cos

approxPi :: Floating t => t

Approximate pi, synonym for Prelude's pi.

We do not define (exect) pi in this package as we have no type that can represent it exactly.

Orphan instances

CanPow Double Double

Associated Types

type PowType Double Double :: *

Methods

pow :: Double -> Double -> PowType Double Double

CanPow Double Rational

Associated Types

type PowType Double Rational :: *

Methods

pow :: Double -> Rational -> PowType Double Rational

CanPow Int Double

Associated Types

type PowType Int Double :: *

Methods

pow :: Int -> Double -> PowType Int Double

CanPow Integer Double

Associated Types

type PowType Integer Double :: *

Methods

pow :: Integer -> Double -> PowType Integer Double

CanPow Rational Double

Associated Types

type PowType Rational Double :: *

Methods

pow :: Rational -> Double -> PowType Rational Double