Numeric.MixedTypes.Round

Description

Rounding operations

class CanRound t where

A replacement for Prelude's RealFrac operations, such as round in which the result type is fixed to Integer.

If RealFrac t and CanTestPosNeg t, then one can use the default implementation to mirror Prelude's round, etc.

In other cases, it is sufficient to define properFraction.

Methods

properFraction :: t -> (Integer, t)

properFraction :: RealFrac t => t -> (Integer, t)

truncate :: t -> Integer

round :: t -> Integer

round :: HasOrderCertainly t Rational => t -> Integer

ceiling :: t -> Integer

ceiling :: CanTestPosNeg t => t -> Integer

floor :: t -> Integer

floor :: CanTestPosNeg t => t -> Integer

class HasIntegerBounds t where

Methods

integerBounds :: t -> (Integer, Integer)

integerBounds :: CanRound t => t -> (Integer, Integer)

Tests

specCanRound :: (CanRoundX t, HasIntegers t) => T t -> Spec

HSpec properties that each implementation of CanRound should satisfy.

specHasIntegerBounds :: HasIntegerBoundsX t => T t -> Spec

HSpec properties that each implementation of CanRound should satisfy.