Control.CollectErrors

Documentation

data CollectErrors es v

A wrapper around values which can accommodate a list of (potential) errors that have (maybe) occurred during the computation of a value. A value may be missing, leaving only the error(s).

Such error collectoin allows one to write expressions with partial functions (ie functions that fail for some inputs) instead of branching after each application of such function. Dealing with the errors can be moved outside the expression. If the error data contain enough information, their list can be used to trace the source of the errors.

Constructors

CollectErrors (Maybe v) es

noErrors :: Monoid es => v -> CollectErrors es v

Wrap a pure value in a CollectErrors record

noValue :: es -> CollectErrors es v

Make CollectErrors record with no value, only errors.

prependErrors :: Monoid es => es -> CollectErrors es v -> CollectErrors es v

Add further errors into a CollectErrors record.

lift2 :: Monoid es => (a -> b -> c) -> CollectErrors es a -> CollectErrors es b -> CollectErrors es c

Add error collection support to a binary function.

unlift2first :: Monoid es => (CollectErrors es a -> CollectErrors es b -> t) -> a -> CollectErrors es b -> t

An utility function for easily defining a binary function with only the second parameter collecting errors, if we have an analogous function that requires both parameters to collect errors.

unlift2second :: Monoid es => (CollectErrors es a -> CollectErrors es b -> t) -> CollectErrors es a -> b -> t

An utility function for easily defining a binary function with only the first parameter collecting errors, if we have an analogous function that requires both parameters to collect errors.

ensureCollectErrors :: CanEnsureCollectErrors es v => v -> EnsureCollectErrors es v

Translate a value of a type a to a value of a type CollectErrors es a except when a already is a CollectErrors type, in which case the value is left as is.

type CanEnsureCollectErrors es v = CanEnsureTypeOp (CollectErrors es) v

type EnsureCollectErrors es v = EnsureTypeOp (CollectErrors es) v

Apply CollectErrors to a type except when the type already is a CollectErrors type.

lift2ensureCE :: (Monoid es, CanEnsureCollectErrors es c) => (a -> b -> c) -> CollectErrors es a -> CollectErrors es b -> EnsureCollectErrors es c

Add error collection support to a binary function whose result may already have collected errors.