Control.CollectErrors

Monad for collecting errors in expressions

data CollectErrors es v

A wrapper around values which can accommodate a list of (potential) errors that have (maybe) occurred during the computation of a value. A value may be missing, leaving only the error(s).

Such error collection allows one to write expressions with partial functions (ie functions that fail for some inputs) instead of branching after each application of such function. Dealing with the errors can be moved outside the expression. If the error data contain enough information, their list can be used to trace the source of the errors.

Constructors

CollectErrors

Fields

getMaybeValueCE :: Maybe v

getErrorsCE :: es

type SuitableForCE es = (Monoid es, Eq es, Show es, CanTestErrorsCertain es)

class CanTestErrorsCertain es where

Minimal complete definition

hasCertainError

Methods

hasCertainError :: es -> Bool

hasCertainErrorCE :: CanTestErrorsCertain es => CollectErrors es v -> Bool

noValueCE :: es -> CollectErrors es v

prependErrorsCE :: Monoid es => es -> CollectErrors es v -> CollectErrors es v

filterValuesWithoutErrorCE :: SuitableForCE es => [CollectErrors es v] -> [v]

getValueIfNoErrorCE :: SuitableForCE es => CollectErrors es v -> (v -> t) -> (es -> t) -> t

A safe way to get a value out of the CollectErrors wrapper.

ce2ConvertResult :: (Typeable t, Show t, SuitableForCE es) => CollectErrors es t -> Either ConvertError t

Tools for avoiding CollectErrors(CollectErrors t) and putting CE inside containers

class (Monoid es, EnsureCE es (EnsureCE es a) ~ EnsureCE es a, EnsureCE es (EnsureNoCE es a) ~ EnsureCE es a, EnsureNoCE es (EnsureCE es a) ~ EnsureNoCE es a, EnsureNoCE es (EnsureNoCE es a) ~ EnsureNoCE es a) => CanEnsureCE es a where

A mechanism for adding and removing CollectErrors to a type in a manner that depends on the shape of the type, especially whether it already has CollectErrors.

Associated Types

type EnsureCE es a

Add CollectErrors to a type except when the type already has CollectErrors in it.

type EnsureNoCE es a

Methods

ensureCE :: Maybe es -> a -> EnsureCE es a

Translate a value of a type a to a value of a type EnsureCE es a.

ensureCE :: EnsureCE es a ~ CollectErrors es a => Maybe es -> a -> EnsureCE es a

Translate a value of a type a to a value of a type EnsureCE es a.

deEnsureCE :: Maybe es -> EnsureCE es a -> Either es a

deEnsureCE :: (EnsureCE es a ~ CollectErrors es a, Eq es) => Maybe es -> EnsureCE es a -> Either es a

ensureNoCE :: Maybe es -> a -> (Maybe (EnsureNoCE es a), es)

ensureNoCE :: (EnsureNoCE es a ~ a, Eq es, Monoid es) => Maybe es -> a -> (Maybe (EnsureNoCE es a), es)

noValueECE :: Maybe a -> es -> EnsureCE es a

Make CollectErrors record with no value, only errors.

noValueECE :: EnsureCE es a ~ CollectErrors es a => Maybe a -> es -> EnsureCE es a

Make CollectErrors record with no value, only errors.

prependErrorsECE :: Maybe a -> es -> EnsureCE es a -> EnsureCE es a

prependErrorsECE :: EnsureCE es a ~ CollectErrors es a => Maybe a -> es -> EnsureCE es a -> EnsureCE es a

getValueOrThrowErrorsNCE

Arguments

:: (SuitableForCE es, CanEnsureCE es v, Show v)

=> Maybe es
sample only

-> v

-> EnsureNoCE es v

An unsafe way to get a value out of an CollectErrors wrapper.

lift1CE :: (SuitableForCE es, CanEnsureCE es a, CanEnsureCE es c) => (a -> c) -> CollectErrors es a -> EnsureCE es c

Add error collection support to an unary function whose result may already have collected errors.

lift2CE :: (SuitableForCE es, CanEnsureCE es a, CanEnsureCE es b, CanEnsureCE es c) => (a -> b -> c) -> CollectErrors es a -> CollectErrors es b -> EnsureCE es c

Add error collection support to a binary function whose result may already have collected errors.

lift2TCE :: (SuitableForCE es, CanEnsureCE es a, CanEnsureCE es c) => (a -> b -> c) -> CollectErrors es a -> b -> EnsureCE es c

Add error collection support to a binary function whose result may already have collected errors. A version where the second operand is not lifted, only the first one.

lift2TLCE :: (SuitableForCE es, CanEnsureCE es b, CanEnsureCE es c) => (a -> b -> c) -> a -> CollectErrors es b -> EnsureCE es c

Add error collection support to a binary function whose result may already have collected errors. A version where the first operand is not lifted, only the second one.

Tools for pulling errors out of structures

class SuitableForCE es => CanExtractCE es f where

Ability to lift collected (potential) errors from inside some structure/collection.

This is useful mostly for structures that use the default implementation of 'CanEnsureCE es'.

Methods

extractCE :: CanEnsureCE es c => Maybe es -> f c -> CollectErrors es (f (EnsureNoCE es c))

extractCE :: (CanEnsureCE es c, Traversable f) => Maybe es -> f c -> CollectErrors es (f (EnsureNoCE es c))