Numeric.MixedTypes.Eq

Description

Equality checks

type HasEq t1 t2 = (HasEqAsymmetric t1 t2, HasEqAsymmetric t2 t1, EqCompareType t1 t2 ~ EqCompareType t2 t1)

class IsBool (EqCompareType a b) => HasEqAsymmetric a b where

Associated Types

type EqCompareType a b

Methods

equalTo :: a -> b -> EqCompareType a b

equalTo :: (EqCompareType a b ~ Bool, a ~ b, Eq a) => a -> b -> Bool

notEqualTo :: a -> b -> EqCompareType a b

notEqualTo :: CanNegSameType (EqCompareType a b) => a -> b -> EqCompareType a b

(==) :: HasEqAsymmetric a b => a -> b -> EqCompareType a b infix 4

(/=) :: HasEqAsymmetric a b => a -> b -> EqCompareType a b infix 4

type HasEqCertainly t1 t2 = (HasEq t1 t2, CanTestCertainly (EqCompareType t1 t2))

type HasEqCertainlyAsymmetric t1 t2 = (HasEqAsymmetric t1 t2, CanTestCertainly (EqCompareType t1 t2))

type HasEqCertainlyCE es t1 t2 = (HasEqCertainly t1 t2, HasEqCertainly (EnsureCE es t1) (EnsureCE es t2))

type HasEqCertainlyCN t1 t2 = HasEqCertainlyCE NumErrors t1 t2

notCertainlyDifferentFrom :: HasEqCertainlyAsymmetric a b => a -> b -> Bool

certainlyEqualTo :: HasEqCertainlyAsymmetric a b => a -> b -> Bool

certainlyNotEqualTo :: HasEqCertainlyAsymmetric a b => a -> b -> Bool

(?==?) :: HasEqCertainlyAsymmetric a b => a -> b -> Bool infix 4

(!==!) :: HasEqCertainlyAsymmetric a b => a -> b -> Bool infix 4

(!/=!) :: HasEqCertainlyAsymmetric a b => a -> b -> Bool infix 4

Tests

specHasEq :: (Show t1, Show t2, Show t3, Arbitrary t1, Arbitrary t2, Arbitrary t3, CanTestCertainly (EqCompareType t1 t1), CanTestCertainly (EqCompareType t1 t2), CanTestCertainly (EqCompareType t2 t1), CanTestCertainly (EqCompareType t2 t3), CanTestCertainly (AndOrType (EqCompareType t1 t2) (EqCompareType t2 t3)), CanAndOrAsymmetric (EqCompareType t1 t2) (EqCompareType t2 t3), HasEqAsymmetric t1 t1, HasEqAsymmetric t1 t2, HasEqAsymmetric t2 t1, HasEqAsymmetric t2 t3) => T t1 -> T t2 -> T t3 -> Spec

HSpec properties that each implementation of HasEq should satisfy.

specHasEqNotMixed :: (Show t, Arbitrary t, CanTestCertainly (EqCompareType t t), CanTestCertainly (AndOrType (EqCompareType t t) (EqCompareType t t)), HasEqAsymmetric t t) => T t -> Spec

HSpec properties that each implementation of HasEq should satisfy.

specConversion :: (Arbitrary t1, Show t1, HasEqCertainly t1 t1) => T t1 -> T t2 -> (t1 -> t2) -> (t2 -> t1) -> Spec

HSpec property of there-and-back conversion.

Specific comparisons

class CanTestNaN t where

Methods

isNaN :: t -> Bool

isNaN :: RealFloat t => t -> Bool

class CanTestFinite t where

Methods

isInfinite :: t -> Bool

isInfinite :: RealFloat t => t -> Bool

isFinite :: t -> Bool

isFinite :: RealFloat t => t -> Bool

class CanTestInteger t where

Minimal complete definition

certainlyNotInteger, certainlyIntegerGetIt

Methods

certainlyNotInteger :: t -> Bool

certainlyInteger :: t -> Bool

certainlyIntegerGetIt :: t -> Maybe Integer

class CanTestZero t where

Methods

isCertainlyZero :: t -> Bool

isCertainlyNonZero :: t -> Bool

isCertainlyZero :: HasEqCertainly t Integer => t -> Bool

isCertainlyNonZero :: HasEqCertainly t Integer => t -> Bool

specCanTestZero :: (CanTestZero t, ConvertibleExactly Integer t) => T t -> Spec

HSpec properties that each implementation of CanTestZero should satisfy.

class CanPickNonZero t where

Methods

pickNonZero :: [(t, s)] -> Maybe (t, s)

Given a list [(a1,b1),(a2,b2),...] and assuming that at least one of a1,a2,... is non-zero, pick one of them and return the corresponding pair (ai,bi).

If none of a1,a2,... is zero, either throws an exception or loops forever.

The default implementation is based on a CanTestZero instance and is not parallel.

pickNonZero :: (CanTestZero t, Show t) => [(t, s)] -> Maybe (t, s)

Given a list [(a1,b1),(a2,b2),...] and assuming that at least one of a1,a2,... is non-zero, pick one of them and return the corresponding pair (ai,bi).

If none of a1,a2,... is zero, either throws an exception or loops forever.

The default implementation is based on a CanTestZero instance and is not parallel.

specCanPickNonZero :: (CanPickNonZero t, CanTestZero t, ConvertibleExactly Integer t, Show t, Arbitrary t) => T t -> Spec

HSpec properties that each implementation of CanPickNonZero should satisfy.